requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
regex==2020.7.14
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
